Update Stress_state_from_csv.py
I have added an import to grab a single data point form the csv file. Ideally this can become a real time frequency of the brain wave. The while loop should run until a keyboard interrupt is made or the loop hits the if i=16 statement(end of file). This statement will have to be changed once we have updating data. Hoping to next add a pyplot with updating display for each iteration of the loop.  The file path will need to be corrected, I just have it set for local use.